move cards into module, import them all for battle sim (#37)
* move cards into module, import them all for battling cards

* fix imports
