+++ kaori/plugins/gacha/engine/test/cards/balanced_cards.py
@@ -1,4 +1,4 @@
-from ..core import *
+from ...core import *
 
 balanced_S = Card(
     name="Balanced S",
