--- kaori/plugins/gacha/engine/test/interesting_cards.py
@@ -1,1 +1,0 @@
-
